Plugin automation fixes from torbenh.
git-svn-id: svn://localhost/ardour2/branches/3.0@3826 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/meter.cc
@@ -37,7 +37,7 @@ PeakMeter::run_in_place (BufferSet& bufs, nframes_t start_frame, nframes_t end_f
 {
 	uint32_t n = 0;
 	uint32_t meterable = std::min(bufs.count().n_total(), (uint32_t)_peak_power.size());
-	uint32_t limit = std::min (meterable, bufs.count().n_midi());
+	uint32_t limit = std::min (meterable, (uint32_t)bufs.count().n_midi());
 
 	// Meter what we have (midi)
 	for ( ; n < limit; ++n) {
